Work around rocket launcher UI engine crash
Hud_SetText() is called from the engine but it crashes in Rui_Draw() as aparently text can no longer be not pinned to anything when this is called. Everything but the text works on the rocket launcher now, text probably needs the vgui_pilot_launcher.res file to be reworked to work with the new changes of the engine.